+++ src/light.rs
@@ -4,11 +4,11 @@ use nalgebra::na::Vec3;
 // Implemente other types of lights
 pub struct Light<V> {
     pos:   V,
-    color: Vec3<f64>
+    color: Vec3<f32>
 }
 
 impl<V> Light<V> {
-    pub fn new(pos: V, color: Vec3<f64>) -> Light<V> {
+    pub fn new(pos: V, color: Vec3<f32>) -> Light<V> {
         Light {
             pos:   pos,
             color: color
